# backend/docs/swagger.yaml
components:
  schemas:
...
    data:
      properties:
        data:
          properties:
            message:
              type: string
          type: object
      type: object
...
    response.StandardResponse:
      allOf:
      - $ref: '#/components/schemas/data'
      properties:
        code:
          type: integer
        data: {}
        message:
          type: string
        sourceId:
          type: string
        traceId:
          type: string
      type: object
...
    worldview.CreateCategoryRequest:
      properties:
        name:
          type: string
      required:
      - name
      type: object
    worldview.CreateItemRequest:
      properties:
        category_id:
          type: integer
        cover_image_url:
          type: string
        description:
          type: string
        name:
          type: string
      required:
      - category_id
      - name
      type: object
...
    worldview.UpdateCategoryRequest:
      properties:
        name:
          type: string
      required:
      - name
      type: object
...
  securitySchemes:
    BearerAuth:
      in: header
      name: Authorization
      type: apiKey
...
info:
  description: This is the API for Test.
  title: API 1.0
  version: "1.0"
openapi: 3.1.0
paths:
...
  /worldview/categories:
    get:
      description: Get all worldview categories for the current user
      parameters:
      - description: Page number
        in: query
        name: page
        schema:
          default: 1
          type: integer
      - description: Number of items per page
        in: query
        name: limit
        schema:
          default: 10
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/data'
                properties:
                  code:
                    type: integer
                  data: {}
                  message:
                    type: string
                  sourceId:
                    type: string
                  traceId:
                    type: string
                type: object
          description: OK
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/response.StandardResponse'
          description: Unauthorized
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/response.StandardResponse'
          description: Failed to fetch categories
      security:
      - BearerAuth: []
      summary: Get all worldview categories
      tags:
      - worldview
    post:
      description: Create a new worldview category for the current user
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/worldview.CreateCategoryRequest'
        description: Create Category Request
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/data'
                properties:
                  code:
                    type: integer
                  data: {}
                  message:
                    type: string
                  sourceId:
                    type: string
                  traceId:
                    type: string
                type: object
          description: Created
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/response.StandardResponse'
          description: Invalid request body
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/response.StandardResponse'
          description: Unauthorized
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/response.StandardResponse'
          description: Failed to create category
      security:
      - BearerAuth: []
      summary: Create a new worldview category
      tags:
      - worldview
  /worldview/categories/{id}:
    delete:
      description: Delete a worldview category by its ID
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        schema:
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/data'
                properties:
                  code:
                    type: integer
                  data: {}
                  message:
                    type: string
                  sourceId:
                    type: string
                  traceId:
                    type: string
                type: object
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/response.StandardResponse'
          description: Invalid ID
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/response.StandardResponse'
          description: Access denied
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/response.StandardResponse'
          description: Category not found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/response.StandardResponse'
          description: Failed to delete category
      security:
      - BearerAuth: []
      summary: Delete a worldview category
      tags:
      - worldview
    get:
      description: Get a single worldview category by its ID
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        schema:
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/data'
                properties:
                  code:
                    type: integer
                  data: {}
                  message:
                    type: string
                  sourceId:
                    type: string
                  traceId:
                    type: string
                type: object
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/response.StandardResponse'
          description: Invalid ID
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/response.StandardResponse'
          description: Access denied
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/response.StandardResponse'
          description: Category not found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/response.StandardResponse'
          description: Failed to get category
      security:
      - BearerAuth: []
      summary: Get a single worldview category
      tags:
      - worldview
    put:
      description: Update a worldview category with the given details
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        schema:
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/worldview.UpdateCategoryRequest'
        description: Update Category Request
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/data'
                properties:
                  code:
                    type: integer
                  data: {}
                  message:
                    type: string
                  sourceId:
                    type: string
                  traceId:
                    type: string
                type: object
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/response.StandardResponse'
          description: Invalid ID or request body
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/response.StandardResponse'
          description: Access denied
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/response.StandardResponse'
          description: Category not found
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/response.StandardResponse'
          description: Failed to update category
      security:
      - BearerAuth: []
      summary: Update a worldview category
      tags:
      - worldview
  /worldview/items:
    get:
      description: Get all worldview items for a given category
      parameters:
      - description: Category ID
        in: query
        name: category_id
        required: true
        schema:
          type: integer
      - description: Page number
        in: query
        name: page
        schema:
          default: 1
          type: integer
      - description: Number of items per page
        in: query
        name: limit
        schema:
          default: 10
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/data'
                properties:
                  code:
                    type: integer
                  data: {}
                  message:
                    type: string
                  sourceId:
                    type: string
                  traceId:
                    type: string
                type: object
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/response.StandardResponse'
          description: category_id is required or invalid
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/response.StandardResponse'
          description: Failed to fetch items
      security:
      - BearerAuth: []
      summary: Get all worldview items for a category
      tags:
      - worldview
    post:
      description: Create a new worldview item for the current user
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/worldview.CreateItemRequest'
        description: Create Item Request
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                allOf:
                - $ref: '#/components/schemas/data'
                properties:
                  code:
                    type: integer
                  data: {}
                  message:
                    type: string
                  sourceId:
                    type: string
                  traceId:
                    type: string
                type: object
          description: Created
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/response.StandardResponse'
          description: Invalid request body
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/response.StandardResponse'
          description: Unauthorized
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/response.StandardResponse'
          description: Failed to create item
      security:
      - BearerAuth: []
      summary: Create a new worldview item
      tags:
      - worldview
...